refactor: use macros for channels
